pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "texas-mushrooms"
version = "0.1.0"
description = "A polite web scraper and data pipeline for texasmushrooms.org"
authors = [{name = "Noah", email = "noah@example.com"}]
requires-python = ">=3.12,<3.14"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "seaborn (>=0.13.2,<0.14.0)",
    "plotly-express (>=0.4.1,<0.5.0)",
    "nbformat (>=5.10.4,<6.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "h3>=3.7.6",
    "pymc>=5.10.0",
    "arviz>=0.17.0",
    "geopandas>=0.14.0",
    "rasterio>=1.3.0",
    "shapely>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
    "types-requests",
    "types-beautifulsoup4",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.poetry.group.dev.dependencies]
ipykernel = "^7.1.0"
pytest = "^9.0.1"

